backend/requirements.txt
Django==2.2.27
django-cors-headers==3.5.0
djangorestframework==3.12.4
Pillow==7.2.0
pytz==2021.3
sqlparse==0.4.2
